Move enums to top-level
So that tooling API doesn't have to refer to BuildLifecycleAwareVirtualFileSystem, which would increase the JAR size significantly.
