+++ subprojects/file-watching/src/main/java/org/gradle/internal/watch/vfs/WatchMode.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2021 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.internal.watch.vfs;
+
+public enum WatchMode {
+    ENABLED(true, "enabled"),
+    DEFAULT(true, "enabled if available"),
+    DISABLED(false, "disabled");
+
+    private final boolean enabled;
+    private final String description;
+
+    WatchMode(boolean enabled, String description) {
+        this.enabled = enabled;
+        this.description = description;
+    }
+
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+}
